Refactored. Tag:29.11.19-17.50
