+++ components/org.wso2.carbon.identity.oauth.ciba/src/main/java/org/wso2/carbon/identity/oauth/ciba/api/CibaAuthServiceImpl.java
@@ -18,13 +18,25 @@
 
 package org.wso2.carbon.identity.oauth.ciba.api;
 
+import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.wso2.carbon.identity.oauth.ciba.common.AuthReqStatus;
+import org.wso2.carbon.identity.oauth.ciba.common.CibaConstants;
+import org.wso2.carbon.identity.oauth.ciba.dao.CibaDAOFactory;
 import org.wso2.carbon.identity.oauth.ciba.dto.CibaAuthRequestDTO;
 import org.wso2.carbon.identity.oauth.ciba.dto.CibaAuthResponseDTO;
 import org.wso2.carbon.identity.oauth.ciba.exceptions.CibaCoreException;
 import org.wso2.carbon.identity.oauth.ciba.model.CibaAuthCodeDO;
-import org.wso2.carbon.identity.oauth.ciba.util.CibaAuthUtil;
+import org.wso2.carbon.identity.oauth.common.exception.InvalidOAuthClientException;
+import org.wso2.carbon.identity.oauth.dao.OAuthAppDO;
+import org.wso2.carbon.identity.oauth2.IdentityOAuth2Exception;
+import org.wso2.carbon.identity.oauth2.util.OAuth2Util;
+
+import java.sql.Timestamp;
+import java.util.Calendar;
+import java.util.TimeZone;
+import java.util.UUID;
 
 /**
  * Provides authentication services.
@@ -36,29 +48,158 @@ public class CibaAuthServiceImpl implements CibaAuthService {
     @Override
     public CibaAuthResponseDTO generateAuthResponseDTO(CibaAuthRequestDTO cibaAuthRequestDTO) throws CibaCoreException {
 
-        if (log.isDebugEnabled()) {
-            log.debug("Creating Authentication Response DTO for the authentication request by the client: " +
-                    cibaAuthRequestDTO.getIssuer());
-        }
-        return processRequest(cibaAuthRequestDTO);
+        // Build DO that is to be persisted.
+        CibaAuthCodeDO cibaAuthCodeDO = generateCibaAuthCodeDO(cibaAuthRequestDTO);
+
+        // Persist DO.
+        persistCibaAuthCode(cibaAuthCodeDO);
+
+        // Return built ResponseDTO.
+        return buildAuthResponseDTO(cibaAuthRequestDTO, cibaAuthCodeDO);
     }
 
     /**
-     * Accepts authentication request DTO, creates DO,persists and respond with responseDTO.
+     * Returns a unique AuthCodeDOKey.
      *
-     * @param cibaAuthRequestDTO CIBA Authentication Request Data Transfer Object.
-     * @return CibaAuthResponseDTO
-     * @throws CibaCoreException
+     * @return String Returns random uuid.
      */
-    private CibaAuthResponseDTO processRequest(CibaAuthRequestDTO cibaAuthRequestDTO) throws CibaCoreException {
+    private static String generateAuthCodeKey() {
 
-        // Build DO that to be persisted.
-        CibaAuthCodeDO cibaAuthCodeDO = CibaAuthUtil.generateCibaAuthCodeDO(cibaAuthRequestDTO);
+        return UUID.randomUUID().toString();
+    }
 
-        // Persist DO.
-        CibaAuthUtil.persistCibaAuthCode(cibaAuthCodeDO);
+    /**
+     * Returns a unique auth_req_id.
+     *
+     * @return String Returns random uuid.
+     */
+    private static String generateAuthRequestId() {
 
-        // Return built ResponseDTO.
-        return CibaAuthUtil.buildAuthResponseDTO(cibaAuthRequestDTO, cibaAuthCodeDO);
+        return UUID.randomUUID().toString();
+    }
+
+    /**
+     * Process and return the expires_in for auth_req_id.
+     *
+     * @param cibaAuthRequestDTO DTO accumulating validated parameters from CibaAuthenticationRequest.
+     * @return long Returns expiry_time of the auth-req_id.
+     */
+    private static long getExpiresIn(CibaAuthRequestDTO cibaAuthRequestDTO) {
+
+        long requestedExpiry = cibaAuthRequestDTO.getRequestedExpiry();
+        if (requestedExpiry == 0) {
+            return CibaConstants.EXPIRES_IN_DEFAULT_VALUE_IN_SEC;
+        } else if (requestedExpiry < CibaConstants.MAXIMUM_REQUESTED_EXPIRY_IN_SEC) {
+            return requestedExpiry;
+        } else {
+            log.warn("(requested_expiry) exceeds default maximum value for the CIBA authentication request made by : " +
+                    cibaAuthRequestDTO.getIssuer());
+            return CibaConstants.MAXIMUM_REQUESTED_EXPIRY_IN_SEC;
+        }
+    }
+
+    /**
+     * Builds and returns Ciba AuthCode DO.
+     *
+     * @param cibaAuthRequestDTO CIBA Request Data Transfer Object.
+     * @return CibaAuthCodeDO.
+     */
+    private static CibaAuthCodeDO generateCibaAuthCodeDO(CibaAuthRequestDTO cibaAuthRequestDTO) {
+
+        CibaAuthCodeDO cibaAuthCodeDO = new CibaAuthCodeDO();
+        long issuedTimeInMillis = Calendar.getInstance(TimeZone.getTimeZone(CibaConstants.UTC)).getTimeInMillis();
+        long lastPolledTimeInMillis = issuedTimeInMillis;
+        Timestamp issuedTime = new Timestamp(issuedTimeInMillis);
+        Timestamp lastPolledTime = new Timestamp(lastPolledTimeInMillis);
+        long expiryTime = getExpiresIn(cibaAuthRequestDTO);
+        String[] scope = cibaAuthRequestDTO.getScope();
+        cibaAuthCodeDO.setCibaAuthCodeKey(generateAuthCodeKey());
+        cibaAuthCodeDO.setAuthReqID(generateAuthRequestId());
+        cibaAuthCodeDO.setConsumerAppKey(cibaAuthRequestDTO.getIssuer());
+        cibaAuthCodeDO.setIssuedTime(issuedTime);
+        cibaAuthCodeDO.setLastPolledTime(lastPolledTime);
+        cibaAuthCodeDO.setAuthenticationStatus(AuthReqStatus.REQUESTED);
+        cibaAuthCodeDO.setInterval(CibaConstants.INTERVAL_DEFAULT_VALUE_IN_SEC);
+        cibaAuthCodeDO.setExpiresIn(expiryTime);
+        cibaAuthCodeDO.setScope(scope);
+        return cibaAuthCodeDO;
+    }
+
+    /**
+     * Builds and returns AuthorizationRequestDTO.
+     *
+     * @param cibaAuthCodeDO     DO with information regarding authenticationRequest.
+     * @param cibaAuthRequestDTO Status of the relevant Ciba Authentication.
+     * @throws CibaCoreException Exception thrown from CibaCore Component.
+     */
+    private static CibaAuthResponseDTO buildAuthResponseDTO(CibaAuthRequestDTO cibaAuthRequestDTO,
+                                                            CibaAuthCodeDO cibaAuthCodeDO) throws CibaCoreException {
+
+        String clientID = cibaAuthRequestDTO.getIssuer();
+        try {
+            CibaAuthResponseDTO cibaAuthResponseDTO = new CibaAuthResponseDTO();
+            String user = cibaAuthRequestDTO.getUserHint();
+            OAuthAppDO appDO = OAuth2Util.getAppInformationByClientId(clientID);
+            String callbackUri = appDO.getCallbackUrl();
+            cibaAuthResponseDTO.setAuthReqId(cibaAuthCodeDO.getAuthReqID());
+            cibaAuthResponseDTO.setCallBackUrl(callbackUri);
+            cibaAuthResponseDTO.setUserHint(user);
+            cibaAuthResponseDTO.setClientId(clientID);
+            cibaAuthResponseDTO.setScopes(OAuth2Util.buildScopeString(cibaAuthRequestDTO.getScope()));
+            cibaAuthResponseDTO.setExpiresIn(cibaAuthCodeDO.getExpiresIn());
+
+            if (StringUtils.isNotBlank(cibaAuthRequestDTO.getBindingMessage())) {
+                cibaAuthResponseDTO.setBindingMessage(cibaAuthRequestDTO.getBindingMessage());
+            }
+
+            if (StringUtils.isNotBlank(cibaAuthRequestDTO.getTransactionContext())) {
+                cibaAuthResponseDTO.setTransactionDetails(cibaAuthRequestDTO.getTransactionContext());
+            }
+
+            if (log.isDebugEnabled()) {
+                log.debug("Successful in creating AuthorizeRequestDTO for the client : " + clientID);
+            }
+            return cibaAuthResponseDTO;
+        } catch (IdentityOAuth2Exception | InvalidOAuthClientException e) {
+            throw new CibaCoreException("Error in creating AuthorizeRequestDTO ", e);
+        }
+    }
+
+    /**
+     * Persist scopes.
+     *
+     * @param cibaAuthCodeDO DO with information regarding authenticationRequest.
+     * @throws CibaCoreException Exception thrown from CibaCore Component.
+     */
+    private static void persistScopes(CibaAuthCodeDO cibaAuthCodeDO) throws CibaCoreException {
+
+        CibaDAOFactory.getInstance().getCibaAuthMgtDAO().storeScope(cibaAuthCodeDO);
+    }
+
+    /**
+     * Persist cibaAuthCode.
+     *
+     * @param cibaAuthCodeDO DO with information regarding authenticationRequest.
+     * @throws CibaCoreException Exception thrown from CibaCore Component.
+     */
+    private static void persistCibaAuthCode(CibaAuthCodeDO cibaAuthCodeDO) throws CibaCoreException {
+
+        CibaDAOFactory.getInstance().getCibaAuthMgtDAO().persistCibaAuthCode(cibaAuthCodeDO);
+        persistScopes(cibaAuthCodeDO);
+    }
+
+    /**
+     * Build and return ACR string as array.
+     *
+     * @param acrString ACR values as a String.
+     * @return String Array.
+     */
+    public static String[] buildACRArray(String acrString) {
+
+        if (StringUtils.isNotBlank(acrString)) {
+            acrString = acrString.trim();
+            return acrString.split("\\s");
+        }
+        return new String[0];
     }
 }
