+++ components/org.wso2.carbon.identity.oauth.ciba/src/test/resources/testng.xml
@@ -20,7 +20,6 @@
 
     <test name="oauth-ciba-test">
         <classes>
-            <class name="org.wso2.carbon.identity.oauth.ciba.util.CibaAuthUtilTest"/>
             <class name="org.wso2.carbon.identity.oauth.ciba.handlers.CibaResponseTypeHandlerTest"></class>
             <class name="org.wso2.carbon.identity.oauth.ciba.grant.CibaGrantHandlerTest"></class>
             <class name="org.wso2.carbon.identity.oauth.ciba.dao.CibaMgtDAOImplTest"></class>
